requirements.txt
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
bidict==0.23.1
cffi==1.17.1
channels==4.1.0
click==8.1.7
constantly==23.10.4
cryptography==43.0.1
daphne==4.1.2
Django==4.2.16
djangorestframework==3.15.2
h11==0.14.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
netifaces==0.10.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyOpenSSL==24.2.1
python-engineio==4.10.1
python-socketio==5.11.4
service-identity==24.1.0
simple-websocket==1.1.0
python-socketio==5.11.4
sqlparse==0.5.1
tomli==2.0.2
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
uvicorn==0.32.0
wsproto==1.2.0
zope.interface==7.1.0
